+++ nextgen/project-builder/tests/project_builder_test.rs
@@ -56,7 +56,6 @@ impl Stub {
             ProjectBuilderContext {
                 detect_language: &self.detect_language,
                 detect_platform: &self.detect_platform,
-                legacy_task_inheritance: false,
                 root_project_id: None,
                 toolchain_config: &self.toolchain_config,
                 workspace_root: &self.workspace_root,