+++ nextgen/project-graph/src/project_graph_builder.rs
@@ -325,10 +325,6 @@ impl<'app> ProjectGraphBuilder<'app> {
             ProjectBuilderContext {
                 detect_language: &context.detect_language,
                 detect_platform: &context.detect_platform,
-                legacy_task_inheritance: !context
-                    .workspace_config
-                    .experiments
-                    .interweaved_task_inheritance,
                 root_project_id: self.root_id.as_ref(),
                 toolchain_config: context.toolchain_config,
                 workspace_root: context.workspace_root,